Moved app creation to app package
