+++ src/website/models.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 """Define database models."""
 
-from . import db
+from app import db
 from flask_login import UserMixin
 from sqlalchemy.sql import func
