+++ src/website/views/auth.py
@@ -3,7 +3,7 @@
 
 from werkzeug.security import generate_password_hash, check_password_hash
 from ..models import User
-from .. import db
+from app import db
 from flask_login import login_user, logout_user, login_required, current_user
 from flask import (
     Blueprint,
