+++ src/website/views/blog.py
@@ -2,7 +2,7 @@
 """Views for the /blog url."""
 
 from ..models import Blogpost
-from .. import db
+from app import db
 from slugify import slugify
 from markdown import markdown
 from flask_login import current_user, login_required
